extract entity + system widgets
