+++ editor/src/main/kotlin/dev/kkorolyov/pancake/editor/widget/GLDetails.kt
