+++ editor/src/main/kotlin/dev/kkorolyov/pancake/editor/widget/SystemsTable.kt
@@ -2,51 +2,19 @@ package dev.kkorolyov.pancake.editor.widget
 
 import dev.kkorolyov.pancake.editor.Widget
 import dev.kkorolyov.pancake.editor.column
-import dev.kkorolyov.pancake.editor.indented
+import dev.kkorolyov.pancake.editor.factory.getGameSystemWidget
 import dev.kkorolyov.pancake.editor.input
-import dev.kkorolyov.pancake.editor.list
 import dev.kkorolyov.pancake.editor.onDoubleClick
 import dev.kkorolyov.pancake.editor.selectable
+import dev.kkorolyov.pancake.editor.separator
 import dev.kkorolyov.pancake.editor.table
 import dev.kkorolyov.pancake.editor.text
-import dev.kkorolyov.pancake.editor.tree
 import dev.kkorolyov.pancake.platform.GameSystem
-import dev.kkorolyov.pancake.platform.Pipeline
 import imgui.ImGui
 import imgui.flag.ImGuiSelectableFlags
 import imgui.flag.ImGuiTableFlags
-import org.lwjgl.glfw.GLFW.*
+import org.lwjgl.glfw.GLFW
 import kotlin.math.roundToInt
-
-/**
- * Renders overall information for [pipelines].
- */
-class PipelinesTree(private val pipelines: Collection<Pipeline>) : Widget {
-	private val details = mutableMapOf<Pipeline, SystemsTable>()
-
-	override fun invoke() {
-		pipelines.forEachIndexed { i, pipeline ->
-			tree("Pipeline $i") {
-				text("Tick time (ns)")
-				indented {
-					text(pipeline.sampler.value)
-				}
-
-				text("TPS")
-				indented {
-					text((1e9 / pipeline.sampler.value).roundToInt())
-				}
-
-				text("Slowest system")
-				indented {
-					text(pipeline.maxBy { it.sampler.value }::class.simpleName ?: "some hook")
-				}
-
-				details.getOrPut(pipeline) { SystemsTable(pipeline.toList()) }()
-			}
-		}
-	}
-}
 
 /**
  * Renders overall information for [systems].
@@ -74,9 +42,9 @@ class SystemsTable(private val systems: Collection<GameSystem>) : Widget {
 						// hook systems (abstract classes most likely) have no details to show
 						name?.let {
 							selectable(it, ImGuiSelectableFlags.SpanAllColumns or ImGuiSelectableFlags.AllowDoubleClick) {
-								detail = SystemDetails(system)
+								detail = getGameSystemWidget(system)
 
-								onDoubleClick(GLFW_MOUSE_BUTTON_1) {
+								onDoubleClick(GLFW.GLFW_MOUSE_BUTTON_1) {
 									details[it] = { Window(it, detail) }
 								}
 							}
@@ -95,34 +63,9 @@ class SystemsTable(private val systems: Collection<GameSystem>) : Widget {
 			}
 		}
 		if (systems.any { it::class.simpleName == null }) input("show hooks", showHooks) { showHooks = it }
+		separator()
 		detail()
 
 		details()
-	}
-}
-
-/**
- * Renders detailed information for [system].
- */
-class SystemDetails(private val system: GameSystem) : Widget {
-	override fun invoke() {
-		text("Tick time (ns)")
-		indented {
-			text(system.sampler.value)
-		}
-
-		text("TPS")
-		indented {
-			text((1e9 / system.sampler.value).roundToInt())
-		}
-
-		text("Signature")
-		indented {
-			list("##signature") {
-				system.forEach {
-					ImGui.selectable(it.simpleName)
-				}
-			}
-		}
 	}
 }
